+++ dashboard/src/scenes/action/list.js
@@ -23,6 +23,7 @@ import { ActionsSelectorsContext } from '../../contexts/selectors';
 import ActionsCalendar from '../../components/ActionsCalendar';
 import SelectCustom from '../../components/SelectCustom';
 import styled from 'styled-components';
+import ActionName from '../../components/ActionName';
 
 const filterActions = (actions, { page, limit, status, currentTeam, search }) => {
   if (status) actions = actions.filter((a) => a.status === status);
@@ -131,7 +132,11 @@ const List = () => {
                   });
                 },
               },
-              { title: 'Nom', dataKey: 'name' },
+              {
+                title: 'Nom',
+                dataKey: 'name',
+                render: (action) => <ActionName action={action} />,
+              },
               {
                 title: 'Personne suivie',
                 dataKey: 'personName',